Refactor Student#save examples
The most important change in this commit was on the test for saving an object whose data already exists in the database.  In the old example, only a change in the ruby object was tested (i.e., the updating of the first_name attribute); the test didn't check whether or not the change reached the database.  Therefore, students' save methods only had to perform inserts, not inserts or updates.

I also refactored the specs to test for changes rather than have multiple expectations in one example and swapped context for describe.