fix issue with update spec
when a student provided a good implementation of save using insert, the
tests for both ADD and UPDATE both passed. The tests for update should
have failed, because the implementation didn't update the existing
record, instead it inserted a new record (which the test read, and
interpreted as the update having worked).